requirements.txt
textdistance
jinja2
google-cloud-storage
